% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{boxcoxtransForecast}
\alias{boxcoxtransForecast}
\title{Performs Box-cox Transformation on a vector}
\usage{
boxcoxtransForecast(one.col, returnLambda = FALSE)
}
\arguments{
\item{one.col}{Any numeric vector that needs to be transformed}

\item{returnLambda}{A boolean value which returns the optimal lambda value for a vector when se to TRUE}
}
\value{
Returns the optimal lambda value for Box-cox transformation when "returnLambda" is set to TRUE and a Box-cox transformed vector otherwise.
}
\description{
Takes in a vector and transforms it using a Box-cox Power Trasformation
}
