% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{GaussianizeLambertW}
\alias{GaussianizeLambertW}
\title{Performs Lambert Transformation on a vector}
\usage{
GaussianizeLambertW(one.col, return_tau_mat = FALSE)
}
\arguments{
\item{one.col}{Any numeric vector that needs to be transformed}

\item{return_tau_mat}{A boolean value which returns the tau_mat vector when set to TRUE}
}
\value{
Returns the tau_mat vector for Lambert-W transformation when "return_tau_mat" is set to TRUE and a Lambert-W transformed vector otherwise.
}
\description{
Takes in a vector and transforms it using a Lmabert-W Trasformation
}
