% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvTransform.R
\name{InvTransform}
\alias{InvTransform}
\title{Fit function that performs inverse transformations to the dataset}
\usage{
InvTransform(data, trans_fit_model)
}
\arguments{
\item{data}{Transformed dataset}

\item{trans_fit_model}{A model file that captures the details of transformations done to given data. This is returned as a list element by VariableTransform function}
}
\value{
An inverse transformed dataset
}
\description{
Takes in transformed dataset and inverse-transform it based on the model file provided
}
