% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataDistribution.R
\name{isnormal}
\alias{isnormal}
\title{Check for the normality of the data}
\usage{
isnormal(data_vector, returnPValue=FALSE)
}
\arguments{
\item{data_vector}{Any data vector that has to be tested for normality}

\item{returnPValue}{A boolean value which when TRUE returns Shapiro P Value}
}
\value{
\itemize{
\item If returnPValue = TRUE, returns the Shapiro P-Value
\item If returnPValue = FALSE, returns TRUE if the data vector is normally dirstributed and FALSE otherwise
}
}
\description{
Takes in a data vector and checks for the normality of the data using Shapiro Test
}
