% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformFit.R
\name{TransformFit}
\alias{TransformFit}
\title{Fit function for test data transformations}
\usage{
TransformFit(data, trans_fit_model)
}
\arguments{
\item{data}{Test data to be tranformed}

\item{trans_fit_model}{A model file that captures the details of transformations done to train data. This is returned as a list element by VariableTransform function}
}
\value{
A transformed dataset
}
\description{
Takes in test data and fits the data based on the model file provided
}
