% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformFit.R
\name{TurkeyFit}
\alias{TurkeyFit}
\title{Performs a Turkey power tranformation with predefined Lambda}
\usage{
TurkeyFit(x, lambda)
}
\arguments{
\item{x}{A vector which neds to be tranformed}

\item{lambda}{A fixed lambda value for turkey power transformation}
}
\value{
Returns log tranformattion if lambda equals zero and a power transormation otherwise
}
\description{
Takes in a vector and a lambda value to perform Turkey power transfomation
}
