% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{VariableTransform}
\alias{VariableTransform}
\title{Transforms the dataset using the best possible transformation technique}
\usage{
VariableTransform(data, ChooseBestTrans = "Shapiro P Value")
}
\arguments{
\item{data}{Any data frame that has atlest one column with continuous data and that has to transformed into normal form}

\item{ChooseBestTrans}{Considers the best transformation based on one of the three values ("Shapiro P Value", "Pearson P Value", "Min skewness")}
}
\value{
Returns a list of 7 objects:
\describe{
\item{transformed_df}{Tranformed Dataset for all Continuous variables}
\item{scaled_df}{Scaled Dataset for all Continuous variables}
\item{original_dist}{Distribution of the dataset provided}
\item{neg_skew_trans}{Intermediate scores (for all the techniques) for Positive skewed data}
\item{pos_skew_trans}{Intermediate scores (for all the techniques) for Positive skewed data}
\item{trans_fit_model}{A list of model fit file, OrderNorm objects and tau_mat objects}
\item{complete_data}{Combination of transformed continuous data and categorical data}
}
}
\description{
VariableTransform initially pads the data inorder to eliminates all negative and zero values and then categorizes the data into normal, positive skewed and negative skewed based on the skewness score of each column of the given data frame

\strong{Transformation for Positive skewed data:}

Different transformation techiniques considered to transform positive skewed data are as follows:
\itemize{
\item Log Transformation
\item Square Root Transformation
\item Cube Root Transformation
\item Turkey Power Transformation
\item Box-Cox Power Transformation
\item Yeo-Johnson Transformation
\item Order-Norm Transformation
\item Lambert-W Transformation
}
Of all these techniques, Best technique is choosed based on the "ChooseBestTrans" argument provided.Default method will be Shapiro P-Value

\strong{Transformation for Negative skewed data:}

Different transformation techiniques considered to transform negative skewed data are as follows:
\itemize{
\item Square Transformation
\item Cube Transformation
\item Turkey Power Transformation
\item Box-Cox Power Transformation
\item Yeo-Johnson Transformation
\item Order-Norm Transformation
\item Lambert-W Transformation
}
Of all these techniques, Best technique is choosed based on the "ChooseBestTrans" argument provided.Default method will be Shapiro P-Value

\strong{Scaling the dataset:}

After merging all the datasets created (normally distributed data, positive skewd data, neggative skewed data), entire dataset is normalized and stored in an other dataframe.
}
\details{
Takes in a data frame and performs the best possible transformation to each of the columns in the data frame
}
